requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
Jinja2==3.1.3
python-multipart==0.0.9
email-validator==2.1.1
# Testing
pytest==8.1.1
passlib==1.7.4
coverage==7.4.4
# LLM
langchain-community==0.3.9
langchain==0.3.9
langchain-experimental==0.3.4
tiktoken==0.8.0
langchain-openai==0.2.10
langchain-google-genai==2.1.4
# utils
pandas==2.2.3
openpyxl==3.1.5
bcrypt==4.2.1
twilio==9.4.5
#qr generation
qrcode[pil]==8.0
# Tabular
tabulate==0.9.0
#firebase
firebase-admin==6.8.0
# Vectordb
langchain-postgres==0.0.5
